vetiver_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=vetiver_requirements.txt 'C:\Users\C25CAL~1.CUR\AppData\Local\Temp\tmpnz120cxv.in'
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
appdirs==1.4.4
    # via pins
build==1.1.1
    # via pip-tools
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pip-tools
    #   rsconnect-python
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
exceptiongroup==1.2.0
    # via anyio
fastapi==0.110.0
    # via vetiver
fsspec==2024.2.0
    # via pins
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via vetiver
humanize==4.9.0
    # via pins
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==7.0.2
    # via
    #   build
    #   pins
importlib-resources==6.1.3
    # via pins
jinja2==3.1.3
    # via pins
joblib==1.3.2
    # via
    #   pins
    #   scikit-learn
    #   vetiver
markupsafe==2.1.5
    # via jinja2
nest-asyncio==1.6.0
    # via vetiver
numpy==1.26.4
    # via
    #   pandas
    #   scikit-learn
    #   scipy
    #   vetiver
packaging==24.0
    # via
    #   build
    #   plotly
pandas==2.2.1
    # via
    #   pins
    #   vetiver
pins==0.8.4
    # via vetiver
pip-tools==7.4.1
    # via vetiver
plotly==5.19.0
    # via vetiver
pydantic==2.6.3
    # via
    #   fastapi
    #   vetiver
pydantic-core==2.16.3
    # via pydantic
pyjwt==2.8.0
    # via rsconnect-python
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via vetiver
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via pins
requests==2.31.0
    # via
    #   pins
    #   vetiver
rsconnect-python==1.22.0
    # via vetiver
scikit-learn==1.4.1.post1
    # via
    #   -r C:\Users\C25CAL~1.CUR\AppData\Local\Temp\tmpnz120cxv.in
    #   vetiver
scipy==1.12.0
    # via scikit-learn
semver==2.13.0
    # via rsconnect-python
six==1.16.0
    # via
    #   python-dateutil
    #   rsconnect-python
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.36.3
    # via fastapi
tenacity==8.2.3
    # via plotly
threadpoolctl==3.3.0
    # via scikit-learn
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.10.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
uvicorn==0.28.0
    # via vetiver
vetiver==0.2.4
    # via -r C:\Users\C25CAL~1.CUR\AppData\Local\Temp\tmpnz120cxv.in
wheel==0.42.0
    # via pip-tools
xxhash==3.4.1
    # via pins
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
